Check is connected in Publish(), call flowComplete() on token when error
There was no check on being connected in the Publish() call as there was in
Subscribe() etc.

In the case where the client was not connected but a call was made to
Subscribe() etc the token error was set but flowComplete() was not called
which caused a deadlock when attempting to wait on the token as there was
no goroutine running which would ever write to it.