Fix test and remove filestore from tests
I've moved the first 3 actions in internalConnLost() to inside the if for
isConnected() as forceDisconnect() was causing internal functions to exit
with an error due to the closed socket but really the intention was to
be disconnecting.

I've also removed the FileStore from being used by default in tests as it
needs some work.