Rewrite and sync with CSD spec
So as part of syncing up with the CSD MQTTv5 spec I have changed a lot
about how the underlying packet handling works, especially regarding
the properties (in the code currently referenced as IDVP - ID Value
Pairs). Currently there is no enforcement or error returned if values
are provided for an invalid packet type, both in encoding and decoding,
I haven't decided yet whether sure checking should be at the packet
level or the client that will build on top.
The list_test.go is a test case working with the packet data and
assumes there will be an mqtt v5 broker listening on 127.0.0.1:1883