fix(356): return string if read after tls handshake fails
In TLS 1.3 the client is the last one to speak in the handshake, so if it causes an error to occur on the server, it will be returned on the client by the first Read, not by Handshake. For example, that will be the case if the server rejects the client certificate.

Bug: https://github.com/eclipse/paho.mqtt.golang/issues/356
Signed-off-by: Armin Galliker <mc_ghc@yahoo.de>