Changed the way to end ack on order matters disabled.
So I noticed when using orders matters, if some of the
subscribred clients send messages with a Quality of Service (QoS) > 0
with slow callbacks, if you close the connexion before callbacks ends,
m.Ack() will be called after the connexion has been closed, leading to sending
message on a closed channel on file net.go, function 'ackFunc' line 451 for QoS 2
or line 458 for QoS1.

So I moved m.Ack() outside the routine, so clients will be notified by Ack (because paho
has correctly recieved the QoS ACK) it's paho's job, but ACK will send even if the callback
did not finished.