Document token.Wait behaviour during reconnect
When a connection is lost to the broker and later resumed the token
returned to the user will be considered completed and a new token will
be created instead. The result of this is that there is no way for a
user of this library to know whether a message has been delivered in the
presence of connection issues.

Signed-off-by: Linus <linus@wallgren.eu>